+++ packages/handlers/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @mailshot/handlers
 
+## 1.4.0
+
+### Minor Changes
+
+- Add reply forwarding via `forwardRepliesTo` sender config option
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @mailshot/shared@1.4.0
+
 ## 1.3.0
 
 ### Minor Changes